Move shared helper modules to `spdl/autoresearch/_common/` (#1451)
Summary:
Pull Request resolved: https://github.com/facebookresearch/spdl/pull/1451

Move 7 independent utility modules from `spdl/tools/autoresearch/` into `spdl/autoresearch/_common/` (no `__init__.py` — consumers import leaf modules directly):

- `_claude.py` — Claude subprocess invocation and JSON extraction
- `_log.py` — logging setup (file + stderr handlers)
- `_state.py` — JSON config/state I/O and master table operations
- `_supervisor.py` — interactive agent process launchers
- `_scm.py` — Git/Sapling abstraction (detect, commit, goto)
- `_jobs.py` — job launch/monitoring stubs
- `_visualization.py` — progress chart and hypothesis tree rendering

These modules know about concepts like workdirs and coding agents but have no coupling to the specific SPDL optimization workflow.

Update all consumer imports across workflow ops, commands, platform, CLI, and tests.

Reviewed By: kiminoue7

Differential Revision: D105895568
