+++ src/spdl/autoresearch/_common/_jobs.py
