+++ src/spdl/autoresearch/_common/_scm.py
