+++ src/spdl/autoresearch/_common/_state.py
