+++ src/spdl/autoresearch/_common/_supervisor.py
