+++ tools/autoresearch/utils/workflow/analysis_ops.py
@@ -13,6 +13,16 @@ import logging
 from datetime import datetime
 from pathlib import Path
 
+from spdl.autoresearch._common._state import (
+    _append_master_row,
+    _read_master_table,
+    write_state,
+)
+from spdl.autoresearch._common._visualization import (
+    _load_tsv,
+    _plot_hypothesis_tree,
+    _plot_progress,
+)
 from spdl.autoresearch.core import (
     AnalysisResult,
     AutoresearchError,
@@ -20,15 +30,9 @@ from spdl.autoresearch.core import (
     FailurePhase,
     HypothesisNode,
 )
-from spdl.tools.autoresearch.plot_progress import (
-    _load_tsv,
-    _plot_hypothesis_tree,
-    _plot_progress,
-)
 
 from ..platform import AutoresearchPlatform
 from ..platform.agents import _parse_agent_result
-from ..state import _append_master_row, _read_master_table, write_state
 from .common import (
     _compare_value,
     _current_best_metric,
