+++ tools/autoresearch/utils/workflow/store.py
@@ -49,9 +49,9 @@ import os
 import time
 from pathlib import Path
 
+from spdl.autoresearch._common._state import SCHEMA_VERSION, write_state
 from spdl.autoresearch.core import HypothesisNode, TaskSpec
 
-from ..state import SCHEMA_VERSION, write_state
 from .policy import (
     _extract_counter,
     _node_from_spec,
